+++ diesel/src/query_dsl/mod.rs
@@ -2,6 +2,7 @@ mod belonging_to_dsl;
 #[doc(hidden)]
 pub mod boxed_dsl;
 mod count_dsl;
+mod distinct_dsl;
 mod group_by_dsl;
 #[doc(hidden)]
 pub mod limit_dsl;
@@ -19,6 +20,7 @@ mod with_dsl;
 pub use self::belonging_to_dsl::BelongingToDsl;
 pub use self::boxed_dsl::BoxedDsl;
 pub use self::count_dsl::CountDsl;
+pub use self::distinct_dsl::DistinctDsl;
 pub use self::filter_dsl::{FilterDsl, FindDsl};
 #[doc(hidden)]
 pub use self::group_by_dsl::GroupByDsl;